--- index.htm
@@ -1,10 +1,0 @@
-<html>
-<body>
-<form action="http://127.0.0.1:8081/process_post" method="POST">
-First Name: <input type="text" name="first_name">  <br>
- 
-Last Name: <input type="text" name="last_name">
-<input type="submit" value="Submit">
-</form>
-</body>
-</html>
