--- upload.htm
@@ -1,14 +1,0 @@
-<html>
-<head>
-<title>文件上传表单</title>
-</head>
-<body>
-<h3>文件上传：</h3>
-选择一个文件上传: <br />
-<form action="/file_upload" method="post" enctype="multipart/form-data">
-<input type="file" name="image" size="50" />
-<br />
-<input type="submit" value="上传文件" />
-</form>
-</body>
-</html>
